refactor(tests): always override ado app dir in tests
