+++ tests/ado/describe/test_ado_describe.py
@@ -29,7 +29,7 @@ def test_describe_nonexistent_space(
     nonexistent_space_id = "i-do-not-exist"
     result = runner.invoke(
         ado,
-        ["--override-ado-app-dir", tmp_path, "describe", "space", nonexistent_space_id],
+        ["describe", "space", nonexistent_space_id],
     )
     assert result.exit_code == 1
     # Travis CI cannot capture output reliably
@@ -54,9 +54,7 @@ def test_describe_valid_space(
         runner=runner, path=tmp_path, project_context=valid_ado_project_context
     )
 
-    result = runner.invoke(
-        ado, ["--override-ado-app-dir", tmp_path, "describe", "space", pfas_space.uri]
-    )
+    result = runner.invoke(ado, ["describe", "space", pfas_space.uri])
     assert result.exit_code == 0
     # AP: TODO: find something actually meaningful to test
