+++ tests/ado/get/test_ado_get_experiments.py
@@ -124,12 +124,8 @@ def test_get_experiments_plural_alias(
         runner=runner, path=tmp_path, project_context=valid_ado_project_context
     )
 
-    result1 = runner.invoke(
-        ado, ["--override-ado-app-dir", tmp_path, "get", "experiment"]
-    )
-    result2 = runner.invoke(
-        ado, ["--override-ado-app-dir", tmp_path, "get", "experiments"]
-    )
+    result1 = runner.invoke(ado, ["get", "experiment"])
+    result2 = runner.invoke(ado, ["get", "experiments"])
     assert result1.exit_code == 0
     assert result2.exit_code == 0
     # Both should produce same output
