+++ tests/conftest.py
@@ -8,6 +8,7 @@ from collections.abc import Callable, Generator
 import pytest
 import ray
 
+from .fixtures.ado_cli_isolation import *
 from .fixtures.core.datacontainer import *
 from .fixtures.core.samplestore import *
 from .fixtures.core.generators import *
